refactor: split functional groups into separate modules
